dropped validation in proto file

+++ internal/domain/adapters/grpc_response.go
@@ -2,5 +2,5 @@ package adapters
 
 type GrpcResponse interface {
 	GetMessage() string
-	GetStatus() uint32
+	GetStatus() int64
 }
